update make target index
The target is now a two-step instead of a three-step process.
The commandsfile generation and the submission script are all in
the same step.

I found that I had been too verbose where it didn't really matter.
I needed to be verbose on the dependencies for the important output
files, not the two-step process for generating the text file for
the run.